+++ totalRP3/tools/Locale.lua
@@ -1106,152 +1106,27 @@ Your profiles, companions profiles and settings will be temporarily stashed away
 	OPTION_ENABLED_TOAST = "Option enabled",
 	OPTION_DISABLED_TOAST = "Option disabled",
 	---@language Markdown
-	WHATS_NEW_17 = [[
-
-## 1.3.0 - 2018-02-19
-
-### New feature: chat links
-
-We have added a new chat links framework to Total RP 3, allowing you to send links for many features of Total RP 3 to other Total RP 3 users. With this update, you can send people links to your profile, companions profiles, at-first-glance presets and other players' profiles from your directory, simply by Shift-Clicking on a profile in the list or on an at-first-glance slot.
-
-{img:Interface\AddOns\totalRP3\resources\1_3_0_link_insert.tga:368:248}
-
-When clicking a link, you can see a quick preview of what was shared in a tooltip, with action buttons available to open or even import the content. When you create a link, you will be asked if you want to let other people import what you are sharing, so you can send a link to your profile but only let people read it.
-
-![Interface\AddOns\totalRP3\resources\1_3_0_link_previews.tga](369,226)
-
-[Total RP 3: Extended] will be updated in the upcoming weeks to use this new framework to share items, campaigns and more.
-
-### New feature: right-click a player to open their profile
-
-We have added a new feature that allows you to right-click on a player in the game to open their profile in Total RP 3 [Toggle feature](right_click_profile). An icon will appear next to the cursor when the player has a RP profile that can be opened. This feature is disabled by default and can be enabled in the Directory settings. It is always disabled when you are in combat, and you can choose to automatically disable it when you go out of character, or apply a modifier like Shift or Control when clicking.
-
-![Interface\AddOns\totalRP3\resources\1_3_0_cursor_icon.tga](335,225)
+	WHATS_NEW_18 = [[
+## [1.3.4](https://github.com/Ellypse/Total-RP-3/compare/1.3.3.2...1.3.4) - 2018-04-18
 
 ### Added
 
-- Added support for the new allied races (default profile icons and languages).
-- Added new "Last seen" column in the directory, allowing you to sort players by the last time you saw them (the column is revealed when the window is resized to be larger).
-- Added new slash command to open someone's profile. Using `/trp3 open` will open your target's profile, while using `/trp3 open CharacterName` (or `/trp3 open CharacterName-RealmName` for connected servers) will request that player for their profile and open it inside Total RP 3 when fully downloaded.
-- Added option to automatically infer a companion name in an emote when the name is placed inside brackets, enabled by default [Toggle feature](companion_speeches). This means you can now create NPC speeches or emotes where a companion name is used, fully formatted with colors.
-
-![Interface\AddOns\totalRP3\resources\1_3_0_npc_speeches.tga](356,224)
-
-- Added color palettes to the color picker, with basic colors, class colors, item quality colors, resources colors, and a custom palette where you can save custom colors you might want to re-use later.
-
-![Interface\AddOns\totalRP3\resources\1_3_0_colors.tga](356,308)
-
-- Added an option to use the default color picker instead of Total RP 3's color picker in the General settings [Toggle setting](default_color_picker). When this option is enabled, the default color picker or any replacement add-on will be used instead when setting a color. When the option is disabled, you can Shift + Click to open the default color picker instead of Total RP 3's.
-- Added an option to disable chat modifications (custom names, text emotes detection, etc.) while set as out of character [Toggle setting](disable_chat_ooc).
-- The option to crop long texts in the tooltips now also crop long text in the at-first-glance tooltip when shown on the target frame. Only applies to other people's at-first-glance presets. The tooltips inside the profile page are not cropped.
-- Updated list of icons for the icon browser, added 177 icons from patch 7.3.5.
-- Updated list of images for the image browser, added 97 images of the game's zones from the new warboard UI from patch 7.3.5.
-- Updated list of musics for the music browser, added 23 musics from patch 7.3 (Argus).
-- Added tooltip flag to indicate when a player is on a trial account.
-
-![Interface\AddOns\totalRP3\resources\1_3_0_trial_flag.tga](250,145)
-
-### Fixed
-
-- Fixed an issue causing incorrect colors in NPC chat messages when using text emotes (`* *` or `< >`) or out of character texts (`( )` or `(( ))`).
-- Text emotes and out of character texts patterns are no longer parsed in links (like item links or achievement links).
-- Fixed issue preventing from clicking on links in companion profiles.
-- Fixed an issue that prevent at-first-glance information from companion profiles from being updated.
-- Fixed an issue with name formatting when Shift + Clicking player names in chat.
+- Chat links now indicate the size of the content in the tooltip.
+- Added an option in the tooltip settings to use the old OOC indicator (|TInterface\COMMON\Indicator-Red:15:15|t) instead of the |cffff0000[OOC]|r text.
+- Added information text on the profile export window to warn about using advanced text editing software when copy-pasting serialized profile as they can alter the data when replacing special characters like quotes.
 
 ### Modified
 
-- The mature profile filter now lets you set a strength when filtering profiles, which affects how many bad words must be found before a profile is considered as mature [Go to settings](open_mature_filter_settings).
-
-![Interface\AddOns\totalRP3\resources\1_3_0_mature.tga](371,119)
-
-- Profiles that have been flagged as mature are now re-evaluated after 24 hours.
-- The previous custom dictionaries of mature language words that we manually made have been replaced by dictionaries from a crowd sourced source, and are now available for German, Italian, Mexican Spanish, and Russian.
-- Various under the hood fixes and improvements in various places, using new libraries and cleaned up some code.
-- Improvements to the localization system. More pets should be recognized as companions, especially when using non-English clients.
-- We are now using a custom popup dialog when clicking on a link to offer to copy-paste the URL, to avoid issues when other add-ons are messing with the game's default dialog.
-- Minor changes to the way some tooltips are displayed (color picker buttons, the minimap button).
-- The Out of character indicator in the tooltip is now more obvious. Instead of the red icon in the top right corner of the tooltip, an [OOC] text in red is displayed in front of the player's name.
-
-![Interface\AddOns\totalRP3\resources\1_3_0_ooc.tga](260,255)
-
-### Removed
-
-- Removed various advanced options that were causing confusion and would often result in the add-on not working when the user changed the default configuration (automatically adding people to the directory, the broadcast channel name).
-
-]],
-	---@language Markdown
-	WHATS_NEW_17_1 = [[
-## 1.3.1 - 2018-02-27
+- The dashboard has received under-the-hood improvements to improve localization and prepare for future features.
+- Streamlined color picker presets UI to avoid confusions. Only the presets button is now visible, bellow the color preview. The presets dropdown now gives access to saving, renaming and deleting custom presets.
+- Player personality traits now only show the value numbers when the cursor is on the bar.
+- Improved generation of chat links for better efficiency.
+- Updated list of Patreon supporters.
 
 ### Fixed
 
-- Fixed an issue with the cropping system introduced to the at-first-glance tooltips when the Register Tooltips module is disabled.
-- Fixed an issue when sorting profiles in the Directory by Last seen where the profiles would be sorted by the dates alphabetically.
-
-### Added
-
-- Added sound related APIs for future Total RP 3: Extended features.
-
-]],
-	---@language Markdown
-	WHATS_NEW_17_2 = [[
-## [1.3.2](https://github.com/Ellypse/Total-RP-3/compare/1.3.1...1.3.2) - 2018-03-06
-
-This update is brought by Meorawr. Huge thanks to him for his contribution to the Total RP 3 project!
-
-### Added
-
-- Added option to hide the toolbar's title.
-- Added an optional alignment parameter to the image tag. Add `:l` for left or `:r` for right (or `:c` for center, but the image is always centered by default) at the end of an image tag to align it to the left or to the right. Note that due to limitations in the game's engine, the text cannot wrap around the image.
-- You can now reorder your Additional information by dragging their icons.
-- Total RP 3's dice rolls feature now supports modifiers to apply to the total for the roll, for example `/trp3 roll 1d20+10` would add 10 to the result of the roll.
-- You can now copy and paste an at-first-glance slot content into another slot, by right-clicking on an at-first-glance slot to copy its content inside Total RP 3's clipboard and right-click on another slot to paste the content.
-
-### Modified
-
-- Personality traits have been completely reworked. The points now go from 0 to 20 (yes, you can pick the extremes now :P), using a slider to more visually set the values. The representation now uses a status bar that is filled proportionally. You can use custom colors of your choice when using custom personality traits. The system automatically converts to and from old values when viewing or sending profiles with people using an older version of Total RP 3.
-- When rolling a single dice using Total RP 3's dice rolls feature you no longer need to precise `1` in front of the dice. Instead of `/trp3 roll 1d20` you can just use `/trp3 roll d20`.
-- The map scanner for players now displays color dots for players who you have set a relationship with. Players with a relationship are also sorted into groups when displayed inside the map tooltips.
-
-### Fixed
-
-- The map scans button is now disabled when the broadcast channel is not available yet or could not be joined. A tooltip indicates the current state of the process.
-- Fixed an issue where the map icons from map scans would not be visible if the user had the option to disable UI animations enabled.
-- Fixed many typos and grammatical errors in the English localization of the add-on (thanks Kruithne for her help!).
-
-]],
-	---@language Markdown
-	WHATS_NEW_17_3 = [[
-## [1.3.3](https://github.com/Ellypse/Total-RP-3/compare/1.3.2.1...1.3.3) - 2018-03-30
-
-### Added
-
-- Added 16 new musics from patch 7.3.5 (Silithus and Seething Shore musics). (Unfortunately, the music paths for the allied races have not been decrupted yet)
-
-### Modified
-
-- Minor Text Fixes
-- Updated libraries and dependencies
-
-]],
-	---@language Markdown
-	WHATS_NEW_17_4 = [[
-## [1.3.3.1](https://github.com/Ellypse/Total-RP-3/compare/1.3.3...1.3.3.1) - 2018-03-31
-
-### Fixed
-
-- Fixed a library compatibility issue when running both Total RP 3 and Storyline.
-
-]],
-	---@language Markdown
-	WHATS_NEW_17_5 = [[
-## [1.3.3.2](https://github.com/Ellypse/Total-RP-3/compare/1.3.3.1...1.3.3.2) - 2018-03-31
-
-### Fixed
-
-- Fixed a critical issue with localizations that are not English or French.
-
+- Fixed a missing option dependency for the mature filter strength slider so that it is correctly disabled when the mature filter is disable.
+- Fixed an issue with the import/export profile UI when a profile had the `~` character inside a field.
 ]],
 	MORE_MODULES_2 = [[{h2:c}Optional modules{/h2}
 {h3}Total RP 3: Extended{/h3}
@@ -1267,7 +1142,6 @@ The Kui |cff9966ffNameplates|r module adds several Total RP 3 customizations to 
 
 
 ]],
-
 	THANK_YOU_ROLE_AUTHOR = "Author",
 	THANK_YOU_ROLE_CONTRIBUTOR = "Contributor",
 	THANK_YOU_ROLE_COMMUNITY_MANAGER = "Community Manager",