+++ CHANGELOG.md
@@ -1,15 +1,25 @@
 # Changelog
 
-## Unreleased
+## [1.3.4](https://github.com/Ellypse/Total-RP-3/compare/1.3.3.2...1.3.4) - 2018-04-18
 
 ### Added
 
 - Chat links now indicate the size of the content in the tooltip.
+- Added an option in the tooltip settings to use the old OOC indicator (the red dot in the top right corner) instead of the [OOC] red text.
+- Added information text on the profile export window to warn about using advanced text editing software when copy-pasting serialized profile as they can alter the data when replacing special characters like quotes.
 
 ### Modified
 
+- The dashboard has received under-the-hood improvements to improve localization and prepare for future features.
+- Streamlined color picker presets UI to avoid confusions. Only the presets button is now visible, bellow the color preview. The presets dropdown now gives access to saving, renaming and deleting custom presets.
 - Player personality traits now only show the value numbers when the cursor is on the bar.
 - Improved generation of chat links for better efficiency.
+- Updated list of Patreon supporters.
+
+### Fixed
+
+- Fixed a missing option dependency for the mature filter strength slider so that it is correctly disabled when the mature filter is disable.
+- Fixed an issue with the import/export profile UI when a profile had the `~` character inside a field.
 
 ## [1.3.3.2](https://github.com/Ellypse/Total-RP-3/compare/1.3.3.1...1.3.3.2) - 2018-03-31